requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9,<3.0
openpyxl>=3.0,<4.0
pandas
